discord/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
ruff==0.4.8
tweepy==4.14.0
